+++ src/main/java/com/refinedmods/refinedstorage/screen/ItemAmountScreen.java
@@ -87,15 +87,7 @@ public class ItemAmountScreen extends AmountSpecifyingScreen<AmountContainer> {
     }
 
     @Override
-    protected void onOkButtonPressed(boolean shiftDown) {
-        try {
-            int amount = Integer.parseInt(amountField.getText());
-
-            RS.NETWORK_HANDLER.sendToServer(new SetFilterSlotMessage(containerSlot, ItemHandlerHelper.copyStackWithSize(stack, amount)));
-
-            close();
-        } catch (NumberFormatException e) {
-            // NO OP
-        }
+    protected void onValidAmountSave(boolean shiftDown, int amount) {
+        RS.NETWORK_HANDLER.sendToServer(new SetFilterSlotMessage(containerSlot, ItemHandlerHelper.copyStackWithSize(stack, amount)));
     }
 }